fix: stop blog tag saves from exhausting the tag identity
Every post save inserted all of its tag names with ON CONFLICT DO NOTHING,
and PostgreSQL consumes an identity value for each skipped row, so the
smallint tag_id ran out after roughly 32,000 tagged saves. Saves now read
existing names and insert only missing ones, and tags.tag_id and
post_tags.tag_id widen to integer; the down migration refuses to narrow
identifiers beyond the smallint range.

+++ rust-be-template/src/features/blog/repository/records.rs
@@ -175,5 +175,5 @@ pub(super) struct NewTagRecord<'a> {
 #[diesel(table_name = post_tags)]
 pub(super) struct NewPostTagRecord {
     pub post_id: Uuid,
-    pub tag_id: i16,
+    pub tag_id: i32,
 }
